[Hackability Refactor] Move cli related files under torchchat/cli (#1083)
