+++ torchchat/cli/convert_hf_checkpoint.py
