+++ torchchat/usages/server.py
@@ -15,7 +15,7 @@ from typing import Dict, List, Union
 
 import torch
 
-from build.builder import BuilderArgs, TokenizerArgs
+from torchchat.cli.builder import BuilderArgs, TokenizerArgs
 from flask import Flask, request, Response
 from generate import GeneratorArgs
